:recycle: 게스트 페이지 api 폴더를 파일 하나로 통합
