+++ src/pages/guest/index.ts
@@ -1,2 +1,2 @@
 export { default } from './ui/page';
-export { requestChannelInfo } from './api/channel';
+export { requestChannelInfo } from './api';
